pdf_rename/requirements.txt
altgraph==0.17.4
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
packaging==25.0
pdfminer.six==20250327
pdfplumber==0.11.6
pefile==2023.2.7
pillow==11.2.1
pycparser==2.22
pyinstaller==6.14.0
pyinstaller-hooks-contrib==2025.4
pypdfium2==4.30.1
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pywin32-ctypes==0.2.3
setuptools==80.9.0
shiboken6==6.9.1
PyYAML==6.0.2
